+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## [Unreleased]
 
 ### Added
+- Show repository projects
 - Show the 'commit-commented' issue event.
+- Make some issue events clickable.
 
 ### Fixed
 - Use the default browser when available.
+- Some minor UI improvements.
 
 ### Translations
 - Improve translations: DE, PT-BR.